+++ src/components/modules/post/PostActionAside.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { m, useAnimationControls } from 'framer-motion'
+import { m, useAnimationControls } from 'motion/react'
 
 import { useIsMobile } from '~/atoms/hooks'
 import { ThumbsupIcon } from '~/components/icons/thumbs-up'
@@ -61,7 +61,7 @@ const SubscribeButton = () => {
   const { present } = usePresentSubscribeModal(['post_c'])
   return (
     <MotionButtonBase className="flex flex-col space-y-2" onClick={present}>
-      <ActionAsideIcon className="icon-[material-symbols--notifications-active-outline] hover:text-accent" />
+      <ActionAsideIcon className="i-material-symbols-notifications-active-outline hover:text-accent" />
     </MotionButtonBase>
   )
 }
@@ -199,7 +199,7 @@ const ShareButton = () => {
         }
       }}
     >
-      <ActionAsideIcon className="icon-[mingcute--share-forward-line] hover:text-uk-cyan-light" />
+      <ActionAsideIcon className="i-mingcute-share-forward-line hover:text-uk-cyan-light" />
     </MotionButtonBase>
   )
 }