+++ src/providers/root/index.tsx
@@ -1,13 +1,11 @@
 'use client'
 
-import { LazyMotion } from 'framer-motion'
-import { SessionProvider } from 'next-auth/react'
+import { LazyMotion } from 'motion/react'
 import { ThemeProvider } from 'next-themes'
 import type { JSX, PropsWithChildren } from 'react'
 
 import { PeekPortal } from '~/components/modules/peek/PeekPortal'
 import { ModalStackProvider } from '~/components/ui/modal'
-import { API_URL } from '~/constants/env'
 import { useBeforeUnload } from '~/hooks/common/use-before-unload'
 
 import { ProviderComposer } from '../../components/common/ProviderComposer'
@@ -27,14 +25,10 @@ const loadFeatures = () =>
   import('./framer-lazy-feature').then((res) => res.default)
 
 const baseContexts: JSX.Element[] = [
-  // @ts-expect-error
   <ThemeProvider key="themeProvider" />,
   <JotaiStoreProvider key="jotaiStoreProvider" />,
 
   <LazyMotion features={loadFeatures} strict key="framer" />,
-  // @ts-expect-error
-  <SessionProvider basePath={`${API_URL}/auth`} key="sessionProvider" />,
-
   <AuthSessionProvider key="authSessionProvider" />,
 ]